A far too late upload of all my work so far
Main Menu added
Test Room made with all features available within
Player Controller that handles movement, attacks, and death
Enemy Movement that allows turtles to path on navmesh
Projectile script to fire a projectile at a set amount of time
A kill plane to act as a pit for the player
Coins that spawn in as scriptable objects
Pause menu that stops player movement
Animations for the player, turtle, and chest.
Coins rotate from a CoRoutine script.